+++ spec/rails_multitenant_spec.rb
@@ -68,14 +68,26 @@ describe "delegating to GlobalContextRegistry" do
     end
   end
 
-  it "RailsMultitenant.with_admin_registry enables admin mode without removing existing context" do
+  it "RailsMultitenant.with_unscoped_queries does not remove existing context" do
     RailsMultitenant::GlobalContextRegistry[:organization_id] = 'Salsify'
 
-    RailsMultitenant::GlobalContextRegistry.with_admin_registry do
-      expect(RailsMultitenant::GlobalContextRegistry[:admin_registry_enabled]).to be_truthy
+    RailsMultitenant::GlobalContextRegistry.with_unscoped_queries do
+      expect(RailsMultitenant::GlobalContextRegistry[:use_unscoped_queries]).to be_truthy
       expect(RailsMultitenant::GlobalContextRegistry[:organization_id]).to eq('Salsify')
     end
 
-    expect(RailsMultitenant::GlobalContextRegistry[:admin_registry_enabled]).to be_falsey
+    expect(RailsMultitenant::GlobalContextRegistry[:use_unscoped_queries]).to be_falsey
+  end
+
+  it "RailsMultitenant.disabled_scoped_queries and RailsMultitenant.enable_scoped_queries do not impact existing context" do
+    RailsMultitenant::GlobalContextRegistry[:organization_id] = 'Salsify'
+
+    RailsMultitenant::GlobalContextRegistry.disable_scoped_queries
+    expect(RailsMultitenant::GlobalContextRegistry[:use_unscoped_queries]).to be_truthy
+    expect(RailsMultitenant::GlobalContextRegistry[:organization_id]).to eq('Salsify')
+
+    RailsMultitenant::GlobalContextRegistry.enable_scoped_queries
+    expect(RailsMultitenant::GlobalContextRegistry[:use_unscoped_queries]).to be_falsey
+    expect(RailsMultitenant::GlobalContextRegistry[:organization_id]).to eq('Salsify')
   end
 end